pyproject.toml
[project]
name = "customer-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "google-genai>=1.21.1",
  "langchain>=0.3.26",
  "langchain-community>=0.3.26",
  "langchain-google-genai>=2.1.5",
  "matplotlib>=3.10.3",
  "networkx>=3.5",
  "nltk>=3.9.1",
  "numpy>=2.3.0",
  "pandas>=2.3.0",
  "pydantic>=2.11.7",
  "python-dotenv>=1.1.0",
  "pyvis>=0.3.2",
  "scikit-learn>=1.7.0",
  "seaborn>=0.13.2",
  "tqdm>=4.67.1",
  "transformers>=4.52.4",
  "torch>=2.0.0",
  "torch-geometric>=2.4.0",
  "plotly>=5.17.0",
  "dash>=2.14.0",
  "dash-cytoscape>=0.3.0",
]

[dependency-groups]
dev = ["ipywidgets>=8.1.7", "ipykernel>=6.29.5", "pytest>=7.0.0"]
